Acceptance matrix, row by row, in one file
The plan lists 26 scenarios. They were covered in scattered places under names
that did not say which row they were, so nobody could answer 'which rows are
green' without reading every test. Every row now appears exactly once: ten
asserted directly, eleven pointing at the test that covers them by path, and
five recorded as needing a human or hardware with the reason stated. A test
asserts that no row is missing from that accounting - an absent row reads as
a passing one.

Lint now covers tests/ as well, and it immediately found two real things:

-  in the off-axis extraction test. It reads as a
  check and can never fail. The intent was real - extraction succeeding beyond
  spec is a bonus, not a failure - so it is a comment now instead of an
  assertion that lies.
- my own , which passes for a typo in the test as
  readily as for the behaviour under test. Now ValidationError specifically.

+++ tests/support/synthetic_sheet.py
@@ -47,7 +47,7 @@ class TrueMetrics:
 def _font_for_cell(cell: layout.CellBox, font_path: Path):
     from PIL import ImageFont
 
-    return ImageFont.truetype(str(font_path), size=int(round(cell.em_px)))
+    return ImageFont.truetype(str(font_path), size=round(cell.em_px))
 
 
 def font_covers(font_path: Path) -> set[str]:
